Enforced new layout so that the UserDataViewer is no longer rendered every time we move to a new page
