+++ src/pages/create-recipe/CreateRecipes.jsx
@@ -1,0 +1,11 @@
+import React from "react";
+import RecipeCreationForm from "./components/RecipeCreationForm";
+
+const CreateRecipes = () => {
+  return (
+    <div>
+      <RecipeCreationForm />
+    </div>
+  );
+};
+export default CreateRecipes;
