+++ src/pages/search/Search.jsx
@@ -1,8 +1,8 @@
 import React, { useState } from "react";
 import { Row, Col, Container, Spinner } from "reactstrap";
-import MealCard from "../../../components/MealCard.jsx";
-import SearchBox from "./SearchBox.jsx";
-import MealDataManager from "../../../utils/MealDataManager.js";
+import MealCard from "../../components/MealCard.jsx";
+import SearchBox from "./components/SearchBox.jsx";
+import MealDataManager from "../../utils/MealDataManager.js";
 import InfiniteScroll from "react-infinite-scroll-component";
 
 const Search = () => {
@@ -29,11 +29,11 @@ const Search = () => {
       const spoonacularQueryResults =
         await mealDataManager.queryRecipeFromSpoonacular(
           query,
-          searchResults.length,
+          searchResults.length
         );
 
       setSearchResults(
-        searchResults.concat(spoonacularQueryResults.resultsList),
+        searchResults.concat(spoonacularQueryResults.resultsList)
       );
       //spoonacular caps results to 1000
       if (searchResults.length >= numResults || searchResults.length >= 999) {
@@ -41,7 +41,7 @@ const Search = () => {
           "searchResults.length=" +
             searchResults.length +
             " numResults=" +
-            numResults,
+            numResults
         );
         setNumResults(false);
       }
